version 1.1.1
- CHANGED "save new recipe to local storage"

This time I read the patients, doctors, recipe periods from local storage.

Also, I added a function to check the emptiness of the form. If it's empty, then the button "Salvare" is disabled, until every field is filled.

I also moved Login and Recover password views to the Views folder because they are views, not components. And I changed some namings in the index.js and App.vue files, to be in English, not Romanian.