version 1.1.15
- solved 0 and month bug. The problem was that when I had dates like 13/10/2025 it showed 13/010/2025 which is not ok. That happened because I forgot that the month starts from 0.

My computer taught that 13/10/2025 is 13/9/2025 and that's why it added a 0. So to solve this, I added +1 to the month in the condition. That was one problem.

Another problem was that if the month or day or both were 10, the date wasn't showed on the screen. That happened because my conditions were wrong, I had > 10 instead of >= 10.

This is the logic of the format date function:
If the date or month is < 10, then add a 0. Anotherwise don't.