version 1.1.15.1
I forgot to mention in the last update:

I modified the function calculateFuturePrescriptionDate()

Now it simply saves the date as it is, like this: 2025-10-11T00:00:00.000Z. Before I formatted, to be like this 2025-10-11.

The date is implicitly converted to string when I save it. I don't have to convert it explicitly with toString().

Also, I did some complicated css to make a part of the text in my <p> bold, in MoreDetailsRecipeView.vue.

You can find them in the previous versions as divs-paragraph-problem and label-problem.

Now I use <span> inside <p> to change the zone that I want to be bold.

Also modified the size of the buttons, in MoreDetailsRecipeView.vue, to make them more narrow